Thought bubble
For every interactable object, a thought bubble (or some other indicator) pops up above the Player's head. This is used to symbolize "Click to interact with" or "Press 'X' to interact with".

This code is in the super class, thus we will not have to copy and paste this code into every object, but rather it will be automatically inherented from any interactable object that is derived from this superclass.