Implemented Class-Structure Using Fan as an Example
Using inheritence, we've created our first batch of interactable objects.

InteractableObject(Super)--just controls the prompt popping up for all interactable items.
Electronic(InteractableObject)--a mid-tier inheritence level, it simply manages checking whether an electronic is plugged in and on.
Fan(Electronic)--our final subclass, handles all of the attributes and functions specific to the fan--in this instance, controlling a box collider that prevents the player's progress to represent being pushed.

ElectricalOutlet(InteractableObject)--a subclass that is mandatorily attached to all Electronic class items; it can be plugged or unplugged, triggering a boolean which then informs the specific subclass (fan, television, etc) how to proceed.