New Camera System
-Known Major Bug: 1/3 times it index errors and tries to shift the camera the wrong way, cannot figure out the root cause but if we like this system I will continue pouring time into investigating.

How it works:
New Script: miniCamTrigger:
due to needing multiple triggers that have a relative location (rather than the old system, in which we used two seperate triggers that operated on a binary two room toggle), made a minitrigger that invokes a call to the main CameraMoveTrigger which is on their parent object.

The invoke calls relativize left and right, then depending on that the actual camera movement is handled in CameraMoveTrigger.

New function, BackToHatto() snaps camera back to the room hatto is in when invoked.

Hatto is frozen via an invoke and a new MoveOn() MoveOff() function in playermover.

Hatto is unfrozen by another invoke, called when the camera snaps back to hatto.