Fixed ? mark always being there when battery is held.
I left it so the out of battery still shows above the roomba--with the battery in hand it should be the intuition we're going for that clicking the roomba will insert the battery and make it run, yes?

Also fixed size of ? mark for the spray bottle. It's scale is/was: .4, .4, 1 for above hatto, but when drew and I add it to items indiidually (him the bottle, me the battery), it uses .8121, .8121, 2. Just fyi.